requirements.txt
Flask>=2.0
Flask-Login>=0.5
Flask-WTF>=1.0
langchain>=0.3.0
tiktoken>=0.5.0

python-dotenv>=0.19
psycopg2-binary
SQLAlchemy>=1.4

tweepy>=4.10
requests
requests-oauthlib

PyPDF2>=2.0
python-docx>=0.8

sentence-transformers>=2.2
faiss-cpu>=1.7
openai>=1.0

APScheduler>=3.8

cryptography>=3.4
werkzeug>=2.0
